fix resource value type

+++ slime/ray/actor_group.py
@@ -24,7 +24,7 @@ class RayTrainGroup:
         wandb_run_id (str, optional): Weights and biases run id. Defaults to None.
         num_gpus_per_actor (float, optional): Number of gpus allocated for each actor.
             If < 1.0, multiple models can share same gpu. Defaults to 1.
-        resources (Dict[str, int], optional): Custom resources to allocate for each actor.
+        resources (Dict[str, float], optional): Custom resources to allocate for each actor.
             See https://docs.ray.io/en/latest/ray-core/scheduling/resources.html
         num_resources_per_node (int, optional): Number of custom resources to allocate for each node.
             See https://docs.ray.io/en/latest/ray-core/scheduling/resources.html
